Amritha m AR poster
A few months ago, I started looking into different AR frameworks and used Unity and Vuforia to create this AR poster. I spent some time putting together this work with the barbarian 3d charecter where when its scanned in our android phones the charecter comes as a 3d object on our screen in out camera. As we import the vuforia core samples assets, weneed to activate them so they can be used and recogonised by the Ar camera game. If everything is setup properly, a 3d model will appear when u point the image target to the laptop's camera. You can either print the image target or just take a picture of it on your phone and point your phone to your computer's screen.